The patch fixes one bug and a quirk
The bug
--------------
Some of the operations (invert) could lead to a ThreadSafeBitSet with a "dirty state".
The unused bits of the last int would not be all equal to 1, which would impact
cardinality and equals.

The quirk
-------------
Two datasets with different "capacity" with the same integers would be equal
iff their capacity has the same reminder modulo 32.
After the commit, 2 datasets need to have the same capacity to be equal
as a prerequisite.